setup: Use regex to find the version
The previous attempt to read the version string from a file in the
`src/<name>` directory failed with py2 since it requires and additional
`__init__.py` in the `src/` path. Instead of injecting constants and
conditionally importing the module just to read the version string, we
use the approach suggested by the authorities on packaging:
https://packaging.python.org/guides/single-sourcing-package-version/
The actual code is (slightly adapted) from `attrs`:
https://www.attrs.org, https://github.com/python-attrs/attrs.

This change simplifies the `main` part quite a bit at the cost of new
functions and a dependence on the `re` package.