Update README
Expands the title a little to be more descriptive and
includes a short overview (quote from the model page).

Fixes the git links to the C source code.

Includes the `codecov` badge. Might be double counting with
`coveralls` available, but we have set it up so why not use it.

[skip ci]