ci: Enable package building and deploying
Enables automatic binary package (wheel) building via `cbuildwheel` and
deploying to the indexes, testpypi always, and pypi on tags.

Uses and older cibuildwheel version to build the python 3.4 wheels and a
later one for the others from 2.7 to 3.8.

To avoid multiple builds and upload trials, we build only on when
running on the travis python 3.7 setup.