Support output as 4-D `xarray.Dataset`s
Initial support for 4-D `xarray` data set output.
Provides even more convenience compared to NumPy broadcasting, it can
also use the spaceweather package for the geomagnetic and solar indices.

Requires spaceweather (there is a package now) and catches import
errors in case it is not installed and points the user to how to install
all the requirements.

Contains a test case to check that broadcasting to 4-D at least works.