ci: Skip wheel building when not deploying
Uses the `before_deploy` section to build the wheels before deploying.
SO far the wheels where always built regardless if the deploy step was
run or not. Moving the build to `before_deploy` should avoid running
unnecessary builds when the deploy step is skipped, e.g. wrong python
version or not on `master`.

Makes it also a tad easier to maintain.