ci: Build and deploy osx and windows wheels
Uses cibuildwheel for windows and macosx wheels to help people that may
be unable to compile the sources themselves for some reason.

Uses a workspace to deploy the windows wheels with linux (travis'
windows machines may leak secrets). While the macosx wheels are built
and deployed within the same job, their workspace dir doesn't exist on
linux and deploy fails, but deploying with macosx works fine.