ci: Pre-install pandas and pin xarray version
Newer pandas can clash with numpy for some reason, and xarray's dataset
output changed in version 0.17 and messes up the doctests.
Both are only used for testing, so we pre-install pandas to match the
numpy version, and pins the xarray version to the one that still prints
the same format that we have in the doctests.