dataset: Use pandas datetimes where possible
Sticks to `pandas` datetimes as long as possible and also for the time
input for the MSIS function. Thus we are not restricted to give an array
of `datetime.datetime` but can use date-time strings in any supported
format which are a bit more userfriendly.

For compatibility with the MSIS functions, using `np.array()` converts
the `pandas` datetime object to an array of `pandas.Timestamp` objects
so that the MSIS function can access the necessary datetime attributes.
`np.array()` is also needed to make numpy broadcasting with [:, None]
work.