dataset: Fix Py2/3 compatibility
Older python/pandas versions handle the `DatetimeIndex` slightly
differently, `.to_numpy()` seems to be more compatible for this.
We also use that converted array for the spaceweather indices because it
keeps the contents as pandas datetime objects and does not convert it to
numpy datetime64. The latter does not have a datetime compatible
`.floor()` function and does *not* provide the necessary attributes
for the MSIS functions.