CI: Remove `coveralls` from pre-install
The `coveralls` package seems to be no longer maintained and we use the
github action now. It also prevented newer versions of `coverage` to be
installed on some systems because it was incompatible. This in turn
resulted in coverage still being slow on Py312 despite trying to address
that with the environment setting.