CI: Update system path for `coveralls` on macOS
Apparently on macOS, there can be multiple versions of `coveralls`, one
from the github action, and one from a python package. The action gets
confused of which one to pick, most certainly it is the one that does
not work and the macOS jobs fail.

Updates the $PATH environment variable to put the homebrew version
first to fix that.